+++ core/nuts-runtime/src/main/java/net/thevpc/nuts/runtime/standalone/util/reflect/FieldReflectProperty.java
@@ -23,6 +23,9 @@
  */
 package net.thevpc.nuts.runtime.standalone.util.reflect;
 
+import net.thevpc.nuts.util.NutsReflectPropertyDefaultValueStrategy;
+import net.thevpc.nuts.util.NutsReflectType;
+
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
 
@@ -34,7 +37,7 @@ public class FieldReflectProperty extends AbstractReflectProperty {
 
     private Field field;
 
-    public FieldReflectProperty(Field field, Object cleanInstance, ReflectType type,ReflectPropertyDefaultValueStrategy defaultValueStrategy) {
+    public FieldReflectProperty(Field field, Object cleanInstance, NutsReflectType type, NutsReflectPropertyDefaultValueStrategy defaultValueStrategy) {
         this.field = field;
         field.setAccessible(true);
         init(field.getName(),type, cleanInstance, field.getGenericType(),defaultValueStrategy);