+++ core/nuts/src/main/java/net/thevpc/nuts/util/NutsReflectConfiguration.java
@@ -1,7 +1,7 @@
 /**
  * ====================================================================
- * Nuts : Network Updatable Things Service
- * (universal package manager)
+ *            Nuts : Network Updatable Things Service
+ *                  (universal package manager)
  * <br>
  * is a new Open Source Package Manager to help install packages and libraries
  * for runtime execution. Nuts is the ultimate companion for maven (and other
@@ -10,7 +10,7 @@
  * other 'things' . Its based on an extensible architecture to help supporting a
  * large range of sub managers / repositories.
  * <br>
- * <p>
+ *
  * Copyright [2020] [thevpc] Licensed under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -21,31 +21,16 @@
  * governing permissions and limitations under the License.
  * <br> ====================================================================
  */
-package net.thevpc.nuts.runtime.standalone.util.reflect;
-
-import net.thevpc.nuts.*;
-import net.thevpc.nuts.util.NutsNameFormat;
-import net.thevpc.nuts.util.NutsStringUtils;
+package net.thevpc.nuts.util;
 
 /**
  *
  * @author thevpc
+ * @since 0.8.4
  */
-public enum ReflectPropertyAccessStrategy implements NutsEnum {
-    FIELD,
-    METHOD,
-    BOTH;
-    private final String id;
+public interface NutsReflectConfiguration {
 
-    ReflectPropertyAccessStrategy() {
-        this.id = NutsNameFormat.ID_NAME.formatName(name());
-    }
+    NutsReflectPropertyAccessStrategy getAccessStrategy(Class clz);
 
-    public static NutsOptional<ReflectPropertyAccessStrategy> parse(String value) {
-        return NutsStringUtils.parseEnum(value, ReflectPropertyAccessStrategy.class);
-    }
-    @Override
-    public String id() {
-        return id;
-    }
+    NutsReflectPropertyDefaultValueStrategy getDefaultValueStrategy(Class clz);
 }