+++ core/nuts/src/main/java/net/thevpc/nuts/util/NutsReflectPropertyDefaultValueStrategy.java
@@ -1,7 +1,7 @@
 /**
  * ====================================================================
- *            Nuts : Network Updatable Things Service
- *                  (universal package manager)
+ * Nuts : Network Updatable Things Service
+ * (universal package manager)
  * <br>
  * is a new Open Source Package Manager to help install packages and libraries
  * for runtime execution. Nuts is the ultimate companion for maven (and other
@@ -10,7 +10,7 @@
  * other 'things' . Its based on an extensible architecture to help supporting a
  * large range of sub managers / repositories.
  * <br>
- *
+ * <p>
  * Copyright [2020] [thevpc] Licensed under the Apache License, Version 2.0 (the
  * "License"); you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -21,15 +21,31 @@
  * governing permissions and limitations under the License.
  * <br> ====================================================================
  */
-package net.thevpc.nuts.runtime.standalone.util.reflect;
+package net.thevpc.nuts.util;
+
+import net.thevpc.nuts.*;
 
 /**
  *
  * @author thevpc
  */
-public interface ReflectConfiguration {
+public enum NutsReflectPropertyDefaultValueStrategy implements NutsEnum {
+    TYPE_DEFAULT,
+    PROPERTY_DEFAULT,
+    NO_DEFAULT;
+    private final String id;
 
-    ReflectPropertyAccessStrategy getAccessStrategy(Class clz);
+    NutsReflectPropertyDefaultValueStrategy() {
+        this.id = NutsNameFormat.ID_NAME.formatName(name());
+    }
 
-    ReflectPropertyDefaultValueStrategy getDefaultValueStrategy(Class clz);
+    public static NutsOptional<NutsReflectPropertyDefaultValueStrategy> parse(String value) {
+        return NutsStringUtils.parseEnum(value, NutsReflectPropertyDefaultValueStrategy.class);
+    }
+
+
+    @Override
+    public String id() {
+        return id;
+    }
 }